requirements.txt
asgiref==3.8.1
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
django-environ==0.11.2
uuid==1.30
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-extensions==3.2.3
